Move votingQueue to src
