+++ src/modules/dpos/validator/components/voteStatus/index.js
