+++ src/modules/dpos/validator/components/voteStatus/statusMessages.js
