+++ src/modules/dpos/validator/components/voteStatus/styles.css
