+++ src/modules/dpos/validator/components/voteSummary/index.js
