+++ src/modules/dpos/validator/components/voteSummary/styles.css
