+++ src/modules/dpos/validator/components/voteSummary/summary.js
