+++ src/modules/dpos/validator/components/voteSummary/summary.test.js
