+++ src/modules/dpos/validator/components/votingQueue/styles.css
@@ -1,4 +1,4 @@
-@import '../../../../../setup/react/app/mixins.css';
+@import '../../../../../../setup/react/app/mixins.css';
 
 .modal {
   width: 45%;
